move commands to args
