+++ src/main.rs
@@ -1,11 +1,8 @@
 #![cfg(not(tarpaulin_include))]
+use cargo_tarpaulin::args::CargoTarpaulinCli;
+use cargo_tarpaulin::cargo::{rust_flags, rustdoc_flags};
 use cargo_tarpaulin::config::{Color, Config, ConfigWrapper};
-use cargo_tarpaulin::{
-    args::TarpaulinCli,
-    cargo::{rust_flags, rustdoc_flags},
-};
 use cargo_tarpaulin::{run, setup_logging};
-use clap::Parser;
 use std::collections::HashMap;
 use tracing::{info, trace};
 
@@ -26,7 +23,7 @@ fn print_env(seen_rustflags: HashMap<String, Vec<String>>, prefix: &str, default
 }
 
 fn main() -> Result<(), String> {
-    let args = from_args();
+    let args = CargoTarpaulinCli::from_args();
 
     let logging_args = args.config.logging;
     setup_logging(
@@ -55,28 +52,6 @@ fn main() -> Result<(), String> {
     // Since this is the last function we run and don't do any error mitigations (other than
     // printing the error to the user it's fine to unwrap here
     run(&config.0).map_err(|e| e.to_string())
-}
-
-#[derive(Debug, Parser)]
-#[command(name = "cargo-tarpaulin")]
-#[command(bin_name = "cargo")]
-#[command(author, version, about, long_about = None)]
-enum CargoTarpaulinCli {
-    Tarpaulin(TarpaulinCli),
-}
-
-impl CargoTarpaulinCli {
-    pub fn args(self) -> TarpaulinCli {
-        match self {
-            CargoTarpaulinCli::Tarpaulin(args) => args,
-        }
-    }
-}
-
-fn from_args() -> TarpaulinCli {
-    CargoTarpaulinCli::try_parse()
-        .map(CargoTarpaulinCli::args)
-        .unwrap_or_else(|_| TarpaulinCli::parse())
 }
 
 fn print_flags<F>(config: &ConfigWrapper, flags_fn: F, prefix: &str)
